# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API SUNAT - Sistema de Facturación Electrónica",
        "description": "API REST para emisión de comprobantes electrónicos SUNAT (Facturas, Boletas, Notas de Crédito/Débito)",
        "contact": {
            "name": "Soporte Técnico",
            "email": "soporte@tuempresa.com"
        },
        "license": {
            "name": "Privada",
            "url": "https://tuempresa.com/license"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor de Desarrollo Local"
        },
        {
            "url": "http://localhost:8000",
            "description": "Servidor Local Docker"
        }
    ],
    "paths": {
...
        "/api/v1/credit-notes": {
            "get": {
                "tags": [
                    "Notas de Crédito"
                ],
                "summary": "Listar notas de crédito",
                "description": "Obtiene un listado paginado de notas de crédito con filtros",
                "operationId": "c0558304d5c17dd33e5bbdff42223bcf",
                "parameters": [
                    {
                        "name": "company_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "branch_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "estado_sunat",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "tipo_doc_afectado",
                        "in": "query",
                        "schema": {
                            "description": "01=Factura, 03=Boleta",
                            "type": "string"
                        }
                    },
                    {
                        "name": "fecha_desde",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "fecha_hasta",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado de notas de crédito"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Notas de Crédito"
                ],
                "summary": "Crear nota de crédito",
                "description": "Crea una nueva nota de crédito para anular o corregir una factura o boleta",
                "operationId": "1447accac064a035c5987caf07af316e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "company_id",
                                    "branch_id",
                                    "client_id",
                                    "serie",
                                    "tipo_doc_afectado",
                                    "num_doc_afectado",
                                    "tipo_nota_credito",
                                    "motivo_nota",
                                    "items"
                                ],
                                "properties": {
                                    "company_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "branch_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "client_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "serie": {
                                        "description": "Serie para nota de crédito",
                                        "type": "string",
                                        "example": "FC01"
                                    },
                                    "tipo_doc_afectado": {
                                        "description": "01=Factura, 03=Boleta",
                                        "type": "string",
                                        "example": "01"
                                    },
                                    "num_doc_afectado": {
                                        "description": "Serie-Número del documento a afectar",
                                        "type": "string",
                                        "example": "F001-000001"
                                    },
                                    "tipo_nota_credito": {
                                        "description": "Código de motivo SUNAT",
                                        "type": "string",
                                        "example": "01"
                                    },
                                    "motivo_nota": {
                                        "type": "string",
                                        "example": "Anulación de la operación"
                                    },
                                    "moneda": {
                                        "description": "PEN o USD",
                                        "type": "string",
                                        "example": "PEN"
                                    },
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "descripcion": {
                                                    "type": "string",
                                                    "example": "Producto 1"
                                                },
                                                "cantidad": {
                                                    "type": "number",
                                                    "example": 1
                                                },
                                                "precio_unitario": {
                                                    "type": "number",
                                                    "example": 100
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Nota de crédito creada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Nota de crédito creada correctamente"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Errores de validación"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/credit-notes/{id}/send-sunat": {
            "post": {
                "tags": [
                    "Notas de Crédito"
                ],
                "summary": "Enviar nota de crédito a SUNAT",
                "description": "Envía la nota de crédito a SUNAT para su validación. Genera XML y recibe CDR.",
                "operationId": "77805984c7c3ee60d2507764ec411c95",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la nota de crédito",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Nota de crédito enviada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Nota de crédito enviada a SUNAT exitosamente"
                                        },
                                        "estado_sunat": {
                                            "type": "string",
                                            "example": "ACEPTADO"
                                        },
                                        "xml_path": {
                                            "type": "string",
                                            "example": "notas_credito/xml/17112025/FC01-000001.xml"
                                        },
                                        "cdr_path": {
                                            "type": "string",
                                            "example": "notas_credito/cdr/17112025/R-FC01-000001.zip"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Nota de crédito no encontrada"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Ingrese su token de acceso obtenido del endpoint /api/auth/login",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}